refactor: Uri 이미지를 최적화해서 파일로 만드는 코드를 유틸 파일로 분리
